+++ randovania/games/prime1/game_data.py
@@ -154,6 +154,12 @@ game_data: randovania.game.data.GameData = randovania.game.data.GameData(
             "- Storage Cavern\n"
             "- Every room after Twin Fires Tunnel\n",
         ),
+        (
+            "I picked up Gravity Suit too early and it is inhibiting my ability to perform gravityless "
+            "slope jumps. How do I proceed?",
+            "You can temporarily unequip Gravity Suit by performing the following:\n\n"
+            "\n\n```\nStart Menu --> Options --> Visor / Suit --> Gravity Suit --> Off\n```",
+        ),
     ],
     web_info=randovania.game.web_info.GameWebInfo(
         what_can_randomize=[